An S4 class to store the output of fitSmoothHazard

This class inherits from vglm-class.

summary(object, ...)

# S4 method for CompRisk
summary(object)

Arguments

object

Object of class CompRisk

...

Extra parameters

Slots

originalData

Data.frame containing the original data (i.e. before case-base sampling). This is used by the absoluteRisk function.

typeEvents

Numeric factor which encodes the type of events being considered (including censoring).

timeVar

Character string giving the name of the time variable, as appearing in originalData

eventVar

Character string giving the name of the event variable, as appearing in originalData