German Breast Cancer Study Group 2

A data frame containing the observations from the GBSG2 study. This is taken almost verbatim from the TH.data package.

brcancer

Format

This data frame contains the observations of 686 women:

horTh

hormonal therapy, a factor at two levels no and yes.

hormon

numeric version of horTh

age

of the patients in years.

menostat

menopausal status, a factor at two levels pre (premenopausal) and post (postmenopausal).

meno

Numeric version of menostat

tsize

tumor size (in mm).

tgrade

tumor grade, a ordered factor at levels I < II < III.

pnodes

number of positive nodes.

progrec

progesterone receptor (in fmol).

estrec

estrogen receptor (in fmol).

time

recurrence free survival time (in days).

cens

censoring indicator (0- censored, 1- event).

Source

Torsten Hothorn (2019). TH.data: TH's Data Archive. R package version 1.0-10. https://CRAN.R-project.org/package=TH.data

References

M. Schumacher, G. Basert, H. Bojar, K. Huebner, M. Olschewski, W. Sauerbrei, C. Schmoor, C. Beyerle, R.L.A. Neumann and H.F. Rauschecker for the German Breast Cancer Study Group (1994), Randomized \(2\times2\) trial evaluating hormonal treatment and the duration of chemotherapy in node-positive breast cancer patients. Journal of Clinical Oncology, 12, 2086--2093.