Compute confidence intervals for risks

This function uses parametric bootstrap to compute confidence intervals for the risk estimates. Since it relies on MLE theory for the validity of these intervals, this function only works for fit_obj that was fitted using family = "glm" (i.e. the default).

# S3 method for absRiskCB
confint(object, parm, level = 0.95, nboot = 500, ...)

Arguments

object

Output of function absoluteRisk.

parm

Output of function fitSmoothHazard that was used to compute object.

level

The confidence level required.

nboot

The number of bootstrap samples to use.

...

Additional arguments for methods.

Value

If there is only one covariate profile, the function returns a matrix with the time points, the risk estimates, and the confidence intervals. If there are more than one covariate profile, the function returns a list with three components.

Details

If the package progress is available, the function also reports on progress of the sampling (which can take some time if there are many covariate profiles and/or time points).