Estrogen plus Progestin and the Risk of Coronary Heart Disease (eprchd)

This data was reconstructed from the curves in figure 2 (Manson 2003).Compares placebo to hormone treatment.

eprchd

Format

A dataframe with 16608 observations and 3 variables:

time

Years (continuous)

status

0=censored, 1=event

treatment

placebo, estPro

References

Manson, J. E., Hsia, J., Johnson, K. C., Rossouw, J. E., Assaf, A. R., Lasser, N. L., ... & Strickland, O. L. (2003). Estrogen plus progestin and the risk of coronary heart disease. New England Journal of Medicine, 349(6), 523-534.

Examples

data("eprchd")
fit <- fitSmoothHazard(status ~ time + treatment, data = eprchd)
#> 'time' will be used as the time variable