Study to Understand Prognoses Preferences Outcomes and Risks of Treatment (SUPPORT)

The support datasets track four response variables: hospital death, severe functional disability, hospital costs and time until death, where the patients are followed for up to 5.56 years.

support

Format

A dataframe with 9104 observations and 30 variables after imputation and the removal of response variables like hospital charges, patient ratio of costs to charges and micro-costs. Ordinal variables, namely functional disability and income were removed. Finally, Surrogate activities of daily living were removed due to sparsity. There were 6 other model scores in the data-set and they were removed; only aps and sps were kept.

Age

Stores a double representing age.

death

Death at any time up to NDI date: 31DEC94.

sex

0=female, 1=male.

hospdead

1=dead in hospital, 0= has not died in hospital.

slos

Days from study entry to discharge.

d.time

days of follow-up.

dzgroup

Each level of dzgroup: ARF/MOSF w/Sepsis, COPD, CHF, Cirrhosis, Coma, Colon Cancer, Lung Cancer, MOSF with malignancy.

dzclass

ARF/MOSF, COPD/CHF/Cirrhosis, Coma and cancer disease classes.

num.co

the number of comorbidities.

edu

years of education of patient.

scoma

The SUPPORT coma score based on Glasgow D3.

avtisst

Average TISS, days 3-25.

race

Indicates race. White, black, asian, hispanic or other.

meanbp

Mean Arterial Blood Pressure Day 3.

wblc

White blood cell count on day 3.

hrt

Heart rate day 3.

resp

Respiration Rate day 3.

temp

Temperature, in celcius, on day 3.

pafi

PaO2/(0.01*FiO2) Day 3.

alb

Serum albumin day 3.

bili

Bilirubin Day 3.

crea

Serum creatinine day 3.

sod

Serum sodium day 3.

ph

Serum pH (in arteries) day 3.

glucose

Serum glucose day 3.

bun

BUN day 3.

urine

urine output day 3.

adlp

ADL patient day 3.

adls

ADL surrogate day 3.

adlsc

Imputed ADL calibrated to surrogate, if a surrogate was used for a follow up.

sps

SUPPORT physiology score

aps

Apache III physiology score

Source

Available at the following website: http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/support2csv.zip. note: must unzip and process this data before use.

Details

Data from the first phase of the SUPPORT study. We imputed the data-set using predictive mean matching, which resulted in 474 complete observations.

References

Knaus WA, Harrell FE, Lynn J et al. (1995): The SUPPORT prognostic model: Objective estimates of survival for seriously ill hospitalized adults. Annals of Internal Medicine 122:191-203.

http://biostat.mc.vanderbilt.edu/wiki/Main/SupportDesc

http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/Csupport.html

Examples

data("support")
mod_cb <- casebase::fitSmoothHazard(death ~ .,
                                   time = "d.time", event="death",
                                   data = support,
                                   ratio = 10)