Fit smooth-in-time parametric hazard functions.

Miettinen and Hanley (2009) explained how case-base sampling can be used to estimate smooth-in-time parametric hazard functions. The idea is to sample person-moments, which may or may not correspond to an event, and then fit the hazard using logistic regression.

fitSmoothHazard(
  formula,
  data,
  time,
  family = c("glm", "gam", "gbm", "glmnet"),
  censored.indicator,
  ratio = 100,
  ...
)

fitSmoothHazard.fit(
  x,
  y,
  formula_time,
  time,
  event,
  family = c("glm", "gbm", "glmnet"),
  censored.indicator,
  ratio = 100,
  ...
)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under Details.

data
a data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which fitSmoothHazard is called.

time
a character string giving the name of the time variable. See Details.

family
a character string specifying the family of regression models used to fit the hazard.

censored.indicator
a character string of length 1 indicating which value in event is the censored. This function will use relevel to set censored.indicator as the reference level. This argument is ignored if the event variable is a numeric.

ratio
integer, giving the ratio of the size of the base series to that of the case series. Defaults to 100.

...
Additional parameters passed to fitting functions (e.g. glm, glmnet, gam).

x
Matrix containing covariates.

y
Matrix containing two columns: one corresponding to time, the other to the event type.

formula_time
A formula describing how the hazard depends on time. Defaults to linear.

event
a character string giving the name of the event variable.

Value

An object of glm and lm when there is only one event of interest, or of class CompRisk, which inherits from vglm, for a competing risk analysis. As such, functions like summary, deviance and coefficients give familiar results.

Details

The object data should either be the output of the function sampleCaseBase or the source dataset on which case-base sampling will be performed. In the latter case, it is assumed that data contains the two columns corresponding to the supplied time and event variables. The variable time is used for the sampling the base series, and therefore it should represent the time variable on its original (i.e. non transformed) scale. If time is missing, the function looks for a column named "time" in the data. Note that the event variable is inferred from formula, since it is the left hand side.

For single-event survival analysis, it is possible to fit the hazard function using glmnet, gam, or gbm. The choice of fitting family is controled by the parameter family. The default value is glm, which corresponds to logistic regression. For competing risk analysis, only glm and glmnet are allowed.

We also provide a matrix interface through fitSmoothHazard.fit, which mimics glm.fit and gbm.fit. This is mostly convenient for family = "glmnet", since a formula interface becomes quickly cumbersome as the number of variables increases. In this setting, the matrix y should have two columns and contain the time and event variables (e.g. like the output of survival::Surv). We need this linear function of time in order to perform case-base sampling. Therefore, nonlinear functions of time should be specified as a one-sided formula through the argument formula_time (the left-hand side is always ignored).

Examples

# Simulate censored survival data for two outcome types from exponential distributions
library(data.table)
nobs <- 500
tlim <- 20

# simulation parameters
b1 <- 200
b2 <- 50

# event type 0-censored, 1-event of interest, 2-competing event
# t observed time/endpoint
# z is a binary covariate
DT <- data.table(z = rbinom(nobs, 1, 0.5))
DT[, `:=`(
  "t_event" = rweibull(nobs, 1, b1),
  "t_comp" = rweibull(nobs, 1, b2)
)]

#>      z   t_event    t_comp
#>   1: 1 217.96303 26.933538
#>   2: 1  59.76338 17.216329
#>   3: 0 204.77071  2.056982
#>   4: 0 100.47209 16.732382
#>   5: 0 378.31412 30.889048
#>  ---                      
#> 496: 1 187.57458 59.305107
#> 497: 0 116.17601  5.115265
#> 498: 1 229.51968 15.852303
#> 499: 0  13.11169  5.597849
#> 500: 1 682.02976 12.787557

DT[, `:=`(
  "event" = 1 * (t_event < t_comp) + 2 * (t_event >= t_comp),
  "time" = pmin(t_event, t_comp)
)]

#>      z   t_event    t_comp event      time
#>   1: 1 217.96303 26.933538     2 26.933538
#>   2: 1  59.76338 17.216329     2 17.216329
#>   3: 0 204.77071  2.056982     2  2.056982
#>   4: 0 100.47209 16.732382     2 16.732382
#>   5: 0 378.31412 30.889048     2 30.889048
#>  ---                                      
#> 496: 1 187.57458 59.305107     2 59.305107
#> 497: 0 116.17601  5.115265     2  5.115265
#> 498: 1 229.51968 15.852303     2 15.852303
#> 499: 0  13.11169  5.597849     2  5.597849
#> 500: 1 682.02976 12.787557     2 12.787557

DT[time >= tlim, `:=`("event" = 0, "time" = tlim)]

#>      z   t_event    t_comp event      time
#>   1: 1 217.96303 26.933538     0 20.000000
#>   2: 1  59.76338 17.216329     2 17.216329
#>   3: 0 204.77071  2.056982     2  2.056982
#>   4: 0 100.47209 16.732382     2 16.732382
#>   5: 0 378.31412 30.889048     0 20.000000
#>  ---                                      
#> 496: 1 187.57458 59.305107     0 20.000000
#> 497: 0 116.17601  5.115265     2  5.115265
#> 498: 1 229.51968 15.852303     2 15.852303
#> 499: 0  13.11169  5.597849     2  5.597849
#> 500: 1 682.02976 12.787557     2 12.787557

out_linear <- fitSmoothHazard(event ~ time + z, DT, ratio = 10)

#> 'time' will be used as the time variable

out_log <- fitSmoothHazard(event ~ log(time) + z, DT, ratio = 10)

#> 'time' will be used as the time variable

# Use GAMs
library(mgcv)

#> Loading required package: nlme

#> This is mgcv 1.8-31. For overview type 'help("mgcv-package")'.

DT[event == 2, event := 1]

#>      z   t_event    t_comp event      time
#>   1: 1 217.96303 26.933538     0 20.000000
#>   2: 1  59.76338 17.216329     1 17.216329
#>   3: 0 204.77071  2.056982     1  2.056982
#>   4: 0 100.47209 16.732382     1 16.732382
#>   5: 0 378.31412 30.889048     0 20.000000
#>  ---                                      
#> 496: 1 187.57458 59.305107     0 20.000000
#> 497: 0 116.17601  5.115265     1  5.115265
#> 498: 1 229.51968 15.852303     1 15.852303
#> 499: 0  13.11169  5.597849     1  5.597849
#> 500: 1 682.02976 12.787557     1 12.787557

out_gam <- fitSmoothHazard(event ~ s(time) + z, DT,
                           ratio = 10, family = "gam")

#> 'time' will be used as the time variable