Create case-base dataset for use in fitting parametric hazard functions

This function implements the case-base sampling approach described in Hanley and Miettinen, Int J Biostatistics 2009. It can be used to fit smooth-in-time parametric functions easily, via logistic regression.

sampleCaseBase(
  data,
  time,
  event,
  ratio = 10,
  comprisk = FALSE,
  censored.indicator
)

Arguments

data
a data.frame or data.table containing the source dataset.

time
a character string giving the name of the time variable. See Details.

event
a character string giving the name of the event variable. See Details.

ratio
Integer, giving the ratio of the size of the base series to that of the case series. Defaults to 10.

comprisk
Logical. Indicates whether we have multiple event types and that we want to consider some of them as competing risks.

censored.indicator
a character string of length 1 indicating which value in event is the censored. This function will use relevel to set censored.indicator as the reference level. This argument is ignored if the event variable is a numeric

Value

The function returns a dataset, with the same format as the source dataset, and where each row corresponds to a person-moment sampled from the case or the base series. otherwise)

Details

The base series is sampled using a multinomial scheme: individuals are sampled proportionally to their follow-up time.

It is assumed that data contains the two columns corresponding to the supplied time and event variables. If either the time or event argument is missing, the function looks for columns with appropriate-looking names (see checkArgsTimeEvent).

Examples

# Simulate censored survival data for two outcome types from exponential distributions
library(data.table)
set.seed(12345)
nobs <- 5000
tlim <- 10

# simulation parameters
b1 <- 200
b2 <- 50

# event type 0-censored, 1-event of interest, 2-competing event
# t observed time/endpoint
# z is a binary covariate
DT <- data.table(z=rbinom(nobs, 1, 0.5))
DT[,`:=` ("t_event" = rweibull(nobs, 1, b1),
          "t_comp" = rweibull(nobs, 1, b2))]

#>       z   t_event     t_comp
#>    1: 1 341.97029 70.4637398
#>    2: 1 333.42420 18.5965942
#>    3: 1  63.76479  6.9837416
#>    4: 1 697.18591  0.9472352
#>    5: 0 107.34263 28.1693592
#>   ---                       
#> 4996: 1  97.02565 95.7599864
#> 4997: 0 128.50435  8.7967661
#> 4998: 1 758.52039  1.6249641
#> 4999: 1 199.31868 60.4051872
#> 5000: 0 275.89061 49.0953403

DT[,`:=`("event" = 1 * (t_event < t_comp) + 2 * (t_event >= t_comp),
         "time" = pmin(t_event, t_comp))]

#>       z   t_event     t_comp event       time
#>    1: 1 341.97029 70.4637398     2 70.4637398
#>    2: 1 333.42420 18.5965942     2 18.5965942
#>    3: 1  63.76479  6.9837416     2  6.9837416
#>    4: 1 697.18591  0.9472352     2  0.9472352
#>    5: 0 107.34263 28.1693592     2 28.1693592
#>   ---                                        
#> 4996: 1  97.02565 95.7599864     2 95.7599864
#> 4997: 0 128.50435  8.7967661     2  8.7967661
#> 4998: 1 758.52039  1.6249641     2  1.6249641
#> 4999: 1 199.31868 60.4051872     2 60.4051872
#> 5000: 0 275.89061 49.0953403     2 49.0953403

DT[time >= tlim, `:=`("event" = 0, "time" = tlim)]

#>       z   t_event     t_comp event       time
#>    1: 1 341.97029 70.4637398     0 10.0000000
#>    2: 1 333.42420 18.5965942     0 10.0000000
#>    3: 1  63.76479  6.9837416     2  6.9837416
#>    4: 1 697.18591  0.9472352     2  0.9472352
#>    5: 0 107.34263 28.1693592     0 10.0000000
#>   ---                                        
#> 4996: 1  97.02565 95.7599864     0 10.0000000
#> 4997: 0 128.50435  8.7967661     2  8.7967661
#> 4998: 1 758.52039  1.6249641     2  1.6249641
#> 4999: 1 199.31868 60.4051872     0 10.0000000
#> 5000: 0 275.89061 49.0953403     0 10.0000000

out <- sampleCaseBase(DT, time = "time", event = "event", comprisk = TRUE)